faster chunk folding; 10X performanc
This is the important one;

previously, an array was used to store the delta values. But, this array would grow to effectively duplicate the input size in memory.
Additionally, an array has overhead (get, push). And, only the last 16 values of the array were used.

So, we can inline those 16 values. It's not pretty, but seems 10X faster on a 1kb input size. That's worth it!