Fix non-deterministic behavior for large buffers
There was a weird issue where the result of `SHA1.fromBytes` would be
non-deterministic when many large (200k elements) `Bytes` objects were
allocated at once. The length and sum were equal, the SHA1 result
somehow not.

A fix, for unclear reasons, is to introduce a `Decode.andThen` guarding
the decoder.